fix!: Linking corrections, new tweaks.

+++ include/jeff/jstring.h
@@ -17,7 +17,7 @@ typedef struct _char_ptrs {
 void str_append_nul(char *str);
 jbool compare_two_strings(const char *const s1, const char *const s2);
 jbool compare_strv(char **const argv, const size_t len);
-char **filter_argv(const uint argc, char **argv);
+char **filter_argv(const size_t argc, char **argv);
 
 #ifdef __cplusplus
 }
